currency_converter/requirements.txt
requests>=2.31.0
streamlit>=1.25.0
python-dotenv>=1.0.0
